backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
channels==4.2.2
mysql-connector-python==9.3.0
django-environ==0.12.0
PyJWT==2.8.0
daphne==4.1.0 